Document new asset services in README
