+++ src/test/java/com/db/assetstore/service/AssetServiceTest.java
@@ -8,38 +8,31 @@ import com.db.assetstore.domain.service.cmd.PatchAssetCommand;
 import com.db.assetstore.domain.model.type.AVString;
 import com.db.assetstore.infra.jpa.AssetEntity;
 import com.db.assetstore.infra.jpa.AttributeEntity;
-import com.db.assetstore.infra.config.JsonMapperProvider;
-import com.db.assetstore.infra.jpa.CommandLogEntity;
 import com.db.assetstore.infra.mapper.AssetMapper;
 import com.db.assetstore.infra.mapper.AttributeMapper;
 import com.db.assetstore.infra.repository.AssetRepository;
 import com.db.assetstore.infra.repository.AttributeRepository;
-import com.db.assetstore.infra.repository.CommandLogRepository;
-import com.db.assetstore.infra.service.AssetCommandServiceImpl;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import com.db.assetstore.infra.service.AssetService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
-import org.mockito.ArgumentCaptor;
 
 import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.*;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
-class AssetCommandServiceImplTest {
+class AssetServiceTest {
 
     AssetMapper assetMapper;
     AttributeMapper attributeMapper;
     AssetRepository assetRepo;
     AttributeRepository attributeRepo;
-    CommandLogRepository commandLogRepository;
-    ObjectMapper objectMapper;
 
-    AssetCommandServiceImpl service;
+    AssetService service;
 
     @BeforeEach
     void setUp() {
@@ -47,19 +40,15 @@ class AssetCommandServiceImplTest {
         attributeMapper = mock(AttributeMapper.class, Mockito.withSettings().defaultAnswer(Mockito.CALLS_REAL_METHODS));
         assetRepo = mock(AssetRepository.class);
         attributeRepo = mock(AttributeRepository.class);
-        commandLogRepository = mock(CommandLogRepository.class);
-        objectMapper = new JsonMapperProvider().objectMapper();
-        when(commandLogRepository.save(any())).thenAnswer(inv -> inv.getArgument(0));
-        service = new AssetCommandServiceImpl(assetMapper, attributeMapper, assetRepo, attributeRepo, commandLogRepository, objectMapper);
 
-        // Default behavior for save to echo the entity
+        service = new AssetService(assetMapper, attributeMapper, assetRepo, attributeRepo);
+
         when(assetRepo.save(any())).thenAnswer(inv -> inv.getArgument(0));
         when(attributeRepo.save(any())).thenAnswer(inv -> inv.getArgument(0));
     }
 
     @Test
     void create_whenEntityAlreadyContainsAttributes_savesDirectlyAndReturnsId() {
-        // given command with attributes
         CreateAssetCommand cmd = CreateAssetCommand.builder()
                 .id("a-1").type(AssetType.CRE)
                 .attributes(List.of(new AVString("city", "Warsaw")))
@@ -75,79 +64,41 @@ class AssetCommandServiceImplTest {
         AssetEntity entity = AssetEntity.builder().id("a-1").type(AssetType.CRE).attributes(List.of(attr)).build();
         when(assetMapper.toEntity(any())).thenReturn(entity);
 
-        // when
-        String id = service.create(cmd);
+        CommandResult<String> result = service.create(cmd);
 
-        // then
-        assertEquals("a-1", id);
+        assertEquals("a-1", result.result());
+        assertEquals("a-1", result.assetId());
         verify(assetRepo, times(1)).save(entity);
         verifyNoInteractions(attributeRepo);
-
-        ArgumentCaptor<CommandLogEntity> logCaptor = ArgumentCaptor.forClass(CommandLogEntity.class);
-        verify(commandLogRepository).save(logCaptor.capture());
-        assertEquals("CreateAssetCommand", logCaptor.getValue().getCommandType());
-        assertEquals("a-1", logCaptor.getValue().getAssetId());
     }
 
     @Test
     void create_whenEntityHasNoAttributes_insertsAllOnCreateAndSaves() {
-        // given command with a single attribute
         CreateAssetCommand cmd = CreateAssetCommand.builder()
                 .id("a-2").type(AssetType.CRE)
                 .attributes(List.of(new AVString("city", "Gdansk")))
                 .executedBy("creator")
                 .build();
 
-        // mapper returns entity with empty list -> triggers insertAllOnCreate
-        AssetEntity entity = AssetEntity.builder().id("a-2").type(AssetType.CRE).attributes(new java.util.ArrayList<>()).build();
+        AssetEntity entity = AssetEntity.builder()
+                .id("a-2")
+                .type(AssetType.CRE)
+                .attributes(new java.util.ArrayList<>())
+                .build();
         when(assetMapper.toEntity(any())).thenReturn(entity);
 
-        // attributeMapper.toEntity will be called and we let real default method build AttributeEntity
-        // when
-        String id = service.create(cmd);
+        CommandResult<String> result = service.create(cmd);
 
-        // then
-        assertEquals("a-2", id);
-        // One save of parent entity after children have been attached
+        assertEquals("a-2", result.result());
+        assertEquals("a-2", result.assetId());
         verify(assetRepo, times(1)).save(entity);
-        // AttributeRepository is not used in insertAllOnCreate path
         verify(attributeRepo, never()).save(any());
-        // The attribute should have been attached to the entity
         assertFalse(entity.getAttributes().isEmpty());
         assertEquals("city", entity.getAttributes().get(0).getName());
-
-        ArgumentCaptor<CommandLogEntity> logCaptor = ArgumentCaptor.forClass(CommandLogEntity.class);
-        verify(commandLogRepository).save(logCaptor.capture());
-        assertEquals("CreateAssetCommand", logCaptor.getValue().getCommandType());
-        assertEquals("a-2", logCaptor.getValue().getAssetId());
     }
 
     @Test
-    void execute_withCreateCommand_dispatchesThroughVisitorAndRecordsLog() {
-        CreateAssetCommand cmd = CreateAssetCommand.builder()
-                .id("exec-1")
-                .type(AssetType.CRE)
-                .attributes(List.of(new AVString("city", "Poznan")))
-                .executedBy("tester")
-                .build();
-
-        AssetEntity entity = AssetEntity.builder().id("exec-1").type(AssetType.CRE).attributes(new java.util.ArrayList<>()).build();
-        when(assetMapper.toEntity(any())).thenReturn(entity);
-
-        CommandResult<String> result = service.execute(cmd);
-
-        assertEquals("exec-1", result.result());
-        assertEquals("exec-1", result.assetId());
-        verify(assetRepo).save(entity);
-
-        ArgumentCaptor<CommandLogEntity> logCaptor = ArgumentCaptor.forClass(CommandLogEntity.class);
-        verify(commandLogRepository).save(logCaptor.capture());
-        assertEquals("CreateAssetCommand", logCaptor.getValue().getCommandType());
-        assertEquals("exec-1", logCaptor.getValue().getAssetId());
-    }
-
-    @Test
-    void update_commonFieldsOnly_persistsChanges() {
+    void patch_commonFieldsOnly_persistsChanges() {
         AssetEntity entity = AssetEntity.builder().id("a-3").type(AssetType.CRE).status("ACTIVE").build();
         when(assetRepo.findByIdAndDeleted("a-3", 0)).thenReturn(Optional.of(entity));
 
@@ -157,47 +108,36 @@ class AssetCommandServiceImplTest {
                 .executedBy("updater")
                 .build();
 
-        service.update(cmd);
+        CommandResult<Void> result = service.patch(cmd);
 
+        assertTrue(result.success());
+        assertEquals("a-3", result.assetId());
         assertEquals("INACTIVE", entity.getStatus());
         verify(assetRepo, times(1)).save(entity);
         verify(attributeRepo, never()).save(any());
-
-        ArgumentCaptor<CommandLogEntity> logCaptor = ArgumentCaptor.forClass(CommandLogEntity.class);
-        verify(commandLogRepository).save(logCaptor.capture());
-        assertEquals("PatchAssetCommand", logCaptor.getValue().getCommandType());
-        assertEquals("a-3", logCaptor.getValue().getAssetId());
     }
 
     @Test
-    void update_attributeChanged_updatesAttributeAndSavesIt() {
-        // existing entity with attribute city="Gdansk"
+    void patch_attributeChanged_updatesAttributeAndSavesIt() {
         AssetEntity parent = AssetEntity.builder().id("a-4").type(AssetType.CRE).attributes(new java.util.ArrayList<>()).build();
         AttributeEntity existing = new AttributeEntity(parent, "city", "Gdansk", Instant.now());
         parent.getAttributes().add(existing);
         when(assetRepo.findByIdAndDeleted("a-4", 0)).thenReturn(Optional.of(parent));
 
-        // incoming patch with city="Warsaw"
         PatchAssetCommand cmd = PatchAssetCommand.builder()
                 .assetId("a-4")
                 .attributes(List.of(new AVString("city", "Warsaw")))
                 .executedBy("modifier")
                 .build();
 
-        service.update(cmd);
+        service.patch(cmd);
 
-        // verify changed and saved
         assertEquals("Warsaw", parent.getAttributes().get(0).getValueStr());
         verify(attributeRepo, times(1)).save(parent.getAttributes().get(0));
-
-        ArgumentCaptor<CommandLogEntity> logCaptor = ArgumentCaptor.forClass(CommandLogEntity.class);
-        verify(commandLogRepository).save(logCaptor.capture());
-        assertEquals("PatchAssetCommand", logCaptor.getValue().getCommandType());
-        assertEquals("a-4", logCaptor.getValue().getAssetId());
     }
 
     @Test
-    void update_attributeSameValue_doesNotSaveAttribute() {
+    void patch_attributeSameValue_doesNotSaveAttribute() {
         AssetEntity parent = AssetEntity.builder().id("a-5").type(AssetType.CRE).attributes(new java.util.ArrayList<>()).build();
         AttributeEntity existing = new AttributeEntity(parent, "city", "Warsaw", Instant.now());
         parent.getAttributes().add(existing);
@@ -209,16 +149,10 @@ class AssetCommandServiceImplTest {
                 .executedBy("modifier")
                 .build();
 
-        service.update(cmd);
+        service.patch(cmd);
 
-        // value remains and attributeRepo.save not called
         assertEquals("Warsaw", parent.getAttributes().get(0).getValueStr());
         verify(attributeRepo, never()).save(any());
-
-        ArgumentCaptor<CommandLogEntity> logCaptor = ArgumentCaptor.forClass(CommandLogEntity.class);
-        verify(commandLogRepository).save(logCaptor.capture());
-        assertEquals("PatchAssetCommand", logCaptor.getValue().getCommandType());
-        assertEquals("a-5", logCaptor.getValue().getAssetId());
     }
 
     @Test
@@ -226,14 +160,11 @@ class AssetCommandServiceImplTest {
         AssetEntity entity = AssetEntity.builder().id("a-6").type(AssetType.CRE).deleted(0).build();
         when(assetRepo.findByIdAndDeleted("a-6", 0)).thenReturn(Optional.of(entity));
 
-        service.delete(DeleteAssetCommand.builder().assetId("a-6").executedBy("deleter").build());
+        CommandResult<Void> result = service.delete(DeleteAssetCommand.builder().assetId("a-6").executedBy("deleter").build());
 
+        assertTrue(result.success());
+        assertEquals("a-6", result.assetId());
         assertEquals(1, entity.getDeleted());
         verify(assetRepo, times(1)).save(entity);
-
-        ArgumentCaptor<CommandLogEntity> logCaptor = ArgumentCaptor.forClass(CommandLogEntity.class);
-        verify(commandLogRepository).save(logCaptor.capture());
-        assertEquals("DeleteAssetCommand", logCaptor.getValue().getCommandType());
-        assertEquals("a-6", logCaptor.getValue().getAssetId());
     }
 }
